bump to v1.1.0

+++ js/general.js
@@ -9,7 +9,6 @@ $(document).ready(function() {
 
         // populate server location
         $('#serverLocation').html(studio.getPref("serverPath"));
-
         $('#setServer').click(function() {
             var f = studio.fileSelectDialog();
             if (f) {
@@ -25,18 +24,20 @@ $(document).ready(function() {
 
         $('#resetPreferences').click(function(e){
             e.preventDefault();
-            studio.sendCommand('FactorySettings');
-            studio.setPreferences('codeEditor.colorScheme',null);
-            studio.extension.storage.setItem('refreshPanels', true);
-            studio.sendExtensionWebZoneCommand('wakanda-extension-web-editor', 'IDE.refreshPreferences');
-            studio.sendExtensionWebZoneCommand('Preferences', '(function(){ window.location.reload(); })');
+            if(confirm('All the user preferences will be set to their original values. Are you sure you want to proceed?')) {
+                studio.sendCommand('FactorySettings');
+                studio.setPreferences('codeEditor.colorScheme',null);
+                studio.extension.storage.setItem('refreshPanels', true);
+                studio.sendExtensionWebZoneCommand('wakanda-extension-web-editor', 'IDE.refreshPreferences');
+                window.top.location.reload();
+            }
         });
 
         if(studio.os.isWindows) {
             $('#environment_variables').remove();
         } else {
             $('#checkInstall').click(function(e) {
-                studio.sendCommand('wakanda-extension-mobile-core.checkDependencies');
+                studio.sendCommand('wakanda-extension-trouble-shooting.checkDependencies');
             });
         }
     });
